# Draft: Rework the scheduler to use threads instead of processes