@Cynical-Optimist 2020-10-14
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 5, 2020, 17:18

added 13 commits

<ul><li>d3bbc5d8 - scheduler.py: Remove &#39;Message&#39; notification type, use the messenger</li><li>48bb7c40 - scheduler.py: Remove task-based notifications and use the state</li><li>055cb218 - _state.py: Only use a single place of truth for the start time</li><li>9ca0c574 - scheduler.py: Pass all &#39;retry&#39; operations through the state</li><li>b65fb9ae - _stream.py: Stop using a &#39;SUSPENDED&#39; event to know the state of the scheduler</li><li>1b7fa99b - _stream.py: Stop using a &#39;TERMINATED&#39; event to know the state of the scheduler</li><li>a4e026e8 - _stream.py: Stop using a &#39;RUNNING&#39; event to know the state of the scheduler</li><li>0ac01c40 - scheduler.py: Remove notifications from scheduler to stream</li><li>9aeeafeb - scheduler.py: Remove all usage of notifications</li><li>264bb01f - _signals.py: allow calling signal handler from non-main threads</li><li>695bed5f - _fixtures.py: Only get the normal number of threads at the start of session</li><li>23b282b3 - scheduler.py: Use threads instead of processes for jobs</li><li>7ca6e601 - _messenger.py: Make `timed_suspendable` public and use it in job.py</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99261279&start_sha=33641df8650c36ffdef09f5179923a89fb9cc61b)