@Cynical-Optimist 2020-10-14
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 6, 2020, 19:34

added 23 commits

<ul><li>ccfca4f5...e79f4a01 - 10 commits from branch <code>master</code></li><li>715a2e3f - scheduler.py: Remove &#39;Message&#39; notification type, use the messenger</li><li>a064c07b - scheduler.py: Remove task-based notifications and use the state</li><li>5d0eeecd - _state.py: Only use a single place of truth for the start time</li><li>af7c128a - scheduler.py: Pass all &#39;retry&#39; operations through the state</li><li>807e7f68 - _stream.py: Stop using a &#39;SUSPENDED&#39; event to know the state of the scheduler</li><li>77b243a6 - _stream.py: Stop using a &#39;TERMINATED&#39; event to know the state of the scheduler</li><li>f747bfc3 - _stream.py: Stop using a &#39;RUNNING&#39; event to know the state of the scheduler</li><li>1b3bedb6 - scheduler.py: Remove notifications from scheduler to stream</li><li>4eab0bb0 - scheduler.py: Remove all usage of notifications</li><li>1d7b66bf - _signals.py: allow calling signal handler from non-main threads</li><li>f05c5042 - _fixtures.py: Only get the normal number of threads at the start of session</li><li>2926ed23 - scheduler.py: Use threads instead of processes for jobs</li><li>dd3465ff - _messenger.py: Make `timed_suspendable` public and use it in job.py</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99443060&start_sha=ccfca4f505a0e8762508a0fe072570db6413e45a)