@Cynical-Optimist 2020-10-14
In GitLab by [[Gitlab user @BenjaminSchubert]](https://gitlab.com/BenjaminSchubert) on Jul 8, 2020, 13:03

added 2 commits

<ul><li>00659def - job.py: Dont&#39; pass the errors through the queue, we can set it directly</li><li>f26ab9fd - job.py: Completely remove the need for a queue between parent and child jobs</li></ul>

[Compare with previous version](/BuildStream/buildstream/-/merge_requests/1982/diffs?diff_id=99814394&start_sha=82197ded517fd2a37370de58d0c65043cb33084b)